(PC-25501) refactor(app):  move HeroButtonList into generic ui components folder
