+++ src/features/identityCheck/pages/identification/ubble/SelectIDOrigin.tsx
@@ -2,10 +2,10 @@ import React, { FunctionComponent, useEffect } from 'react'
 import { Platform, Text } from 'react-native'
 import styled from 'styled-components/native'
 
-import { HeroButtonList } from 'features/identityCheck/components/HeroButtonList'
 import { PageWithHeader } from 'features/identityCheck/components/layout/PageWithHeader'
 import { SecondButtonList } from 'features/identityCheck/components/SecondButtonList'
 import { analytics } from 'libs/analytics'
+import { HeroButtonList } from 'ui/components/buttons/HeroButtonList'
 import { SeparatorWithText } from 'ui/components/SeparatorWithText'
 import { BicolorEarth } from 'ui/svg/icons/BicolorEarth'
 import { BicolorFrance } from 'ui/svg/icons/BicolorFrance'
