+++ src/features/identityCheck/pages/identification/ubble/SelectIDStatus.tsx
@@ -2,11 +2,11 @@ import React, { FunctionComponent, useEffect } from 'react'
 import { Text } from 'react-native'
 import styled from 'styled-components/native'
 
-import { HeroButtonList } from 'features/identityCheck/components/HeroButtonList'
 import { PageWithHeader } from 'features/identityCheck/components/layout/PageWithHeader'
 import { SecondButtonList } from 'features/identityCheck/components/SecondButtonList'
 import { analytics } from 'libs/analytics'
 import { AccessibilityList } from 'ui/components/accessibility/AccessibilityList'
+import { HeroButtonList } from 'ui/components/buttons/HeroButtonList'
 import { SeparatorWithText } from 'ui/components/SeparatorWithText'
 import { BicolorIdCard } from 'ui/svg/icons/BicolorIdCard'
 import { BicolorLostId } from 'ui/svg/icons/BicolorLostId'
