+++ src/features/identityCheck/pages/identification/ubble/SelectPhoneStatus.web.tsx
@@ -1,9 +1,9 @@
 import React, { FunctionComponent } from 'react'
 import styled from 'styled-components/native'
 
-import { HeroButtonList } from 'features/identityCheck/components/HeroButtonList'
 import { PageWithHeader } from 'features/identityCheck/components/layout/PageWithHeader'
 import { AccessibilityList } from 'ui/components/accessibility/AccessibilityList'
+import { HeroButtonList } from 'ui/components/buttons/HeroButtonList'
 import { BicolorNoPhone } from 'ui/svg/icons/BicolorNoPhone'
 import { BicolorPhonePending } from 'ui/svg/icons/BicolorPhonePending'
 import { BicolorSmartphone } from 'ui/svg/icons/BicolorSmartphone'
