+++ src/features/internal/cheatcodes/pages/AppComponents/AppComponents.tsx
@@ -11,7 +11,6 @@ import { CulturalSurveyProgressBar } from 'features/culturalSurvey/components/Cu
 import { CategoryListModule } from 'features/home/components/modules/categories/CategoryListModule'
 import { VenueTile } from 'features/home/components/modules/venues/VenueTile'
 import { categoryBlockList } from 'features/home/fixtures/categoryBlockList.fixture'
-import { HeroButtonList } from 'features/identityCheck/components/HeroButtonList'
 import { SecondButtonList } from 'features/identityCheck/components/SecondButtonList'
 import { Icons } from 'features/internal/cheatcodes/pages/AppComponents/Icons'
 import { Illustrations } from 'features/internal/cheatcodes/pages/AppComponents/Illustrations'
@@ -39,6 +38,7 @@ import { ButtonTertiaryPrimary } from 'ui/components/buttons/ButtonTertiaryPrima
 import { ButtonTertiaryWhite } from 'ui/components/buttons/ButtonTertiaryWhite'
 import { ButtonWithLinearGradient } from 'ui/components/buttons/buttonWithLinearGradient/ButtonWithLinearGradient'
 import { ExternalLink } from 'ui/components/buttons/externalLink/ExternalLink'
+import { HeroButtonList } from 'ui/components/buttons/HeroButtonList'
 import { ClippedTag } from 'ui/components/ClippedTag'
 import { Emoji } from 'ui/components/Emoji'
 import FilterSwitch from 'ui/components/FilterSwitch'
